### 2024-02-08 #627 @RonaldVisser
dendroNetwork
Submitting Author Name: Ronald M. Visser
Submitting Author Github Handle: @RonaldVisser
Repository:  https://github.com/RonaldVisser/dendroNetwork
Version submitted: 0.5.0
Submission type: Standard

Version accepted: TBD

---

```
Package: dendroNetwork
Title: Create networks of dendrochronological series using pairwise similarity
Version: 0.5.0
    c(person(given ="Ronald",
            family = "Visser",
            email = "r.m.visser@saxion.nl",
    person(given = "Angelino",
            family ="Salentino",
            role = "ctb",
   person(given = "Andy",
            family ="Bunn",
            role = "ctb",
            comment = c(ORCID = "0000-0001-9027-2162")))
Maintainer: Ronald Visser <r.m.visser@saxion.nl>
    dplR (>= 1.7.2),
    igraph,
    reshape2,
    RCy3,
    RColorBrewer,
    foreach,
    doParallel,
    lifecycle,
    grDevices
Description: This package makes it easy to create dendrochronological networks based on the similarity between tree-ring series or chronologies. It includes various functions to compare tree-ring curves building upon dplR. The networks can be used to visualise and understand the relations between tree-ring curves. These networks are also very useful to estimate the provenance or wood-use within a structure/context/site.

```

The software deals with the analysis of dendrochronological data and aims the help to study, understand and visualize patterns in the data using network analyses. The software combines R with Cytoscape.

Dendrochronologist. The scientific application lies in the analyses and visualization of large datasets using networks. The method has been applied before to determine the provenance of wood from archaeological contexts, but can also be used to determine other patterns based on the similarity between tree-ring series or even other time series.

- No, none as far as I know.

NA

I tried to run pkgcheck, but it failed with the following feedback:

Error in vapply(tag_data, function(i) i$keyword, character(1L)) :
  values must be length 1,
 but FUN(X[[12]]) result is length 2

I have absolutely no idea why....

The same error occurs locally and using github action.
