### 2024-02-08 #627 @ldecicco-USGS
I'm guessing the issue is that `RCy3` is only on Bioconductor. The mentions I found in the rOpenSci guide:

https://devguide.ropensci.org/building.html?q=Biocond#pkgdependencies

And:
"If you develop a package depending on or intended for Bioconductor, you might find [biocthis](https://lcolladotor.github.io/biocthis/index.html) relevant."

I'll poke around more and see if I can find other recent submissions that used Bioconductor packages.
