### 2024-02-09 #627 @ldecicco-USGS
Thanks for the submission @RonaldVisser ! Looks like a neat package.

Before we can proceed you'll need to get continuous integration set up. I'd start by trying:
https://usethis.r-lib.org/reference/github_actions.html#use-github-action-check-standard-

That might work - I haven't used Bioconductor packages in awhile, so I'm not 100% sure. If that doesn't work, then you could try:
https://rdrr.io/bioc/biocthis/man/use_bioc_github_action.html

The other "X" items above should be pretty straightforward to address, but let me know if you have any questions.

The error on the build is:

```
Quitting from lines 21-44 [flowchart_workflow] (dendroNetwork_use.Rmd)
Error: processing vignette 'dendroNetwork_use.Rmd' failed with diagnostics:
there is no package called 'DiagrammeR'
--- failed re-building ‘dendroNetwork_use.Rmd’
```

If `DiagrammeR` isn't used in any of the package functions, you'll want to add it to "Suggests" for that vignette to be generated.
