### 2024-02-10 #627 @RonaldVisser
Thanks for the compliment! Thank you for the comments and help!

I have done some updating of various functions, added some tests and the continuous integration (github actions R-CMD / pkgcheck) seems to be working fine. According to pkgcheck it now checks all the boxes :)

Due to the various edits, I released a new release (0.5.1), with changes described in the NEWS-file.
