### 2024-02-19 #627 @ldecicco-USGS
---

I was able to run the example functions from the readme and vignette. It would very helpful to push a compiled `pkgdown` site to a GitHub Page. You can set that up with:
```
usethis::use_pkgdown_github_pages()
```
Setting up a pkgdown site should take care of most of the "Documentation " subtasks. The other sub-category that could be improved is "not just perfunctory" vignettes. Perhaps you could take the contents of the "paper.md" and put that same information in a vignette?

I would also ask that you take a look at the following packages and make sure there's not specific overlap:

https://github.com/AndyBunn/dplR
https://github.com/hanecakr/fellingdateR
https://github.com/ropensci/MtreeRing

There's a good chance I'd be asking authors from those packages to do a review. I do not know enough about tree ring science to understand where the overlap might fall.

---
