### 2024-02-20 #627 @RonaldVisser
Thank you for your helpful comments! More steps to make my package better :)

The paper is there to submit to JOSS in a later stage, but I have added an extra Vignette and I'll think about  developing more Vignettes.

I have created a pkgdown website: [https://ronaldvisser.github.io/dendroNetwork/](https://ronaldvisser.github.io/dendroNetwork/)  and that was and interesting and fun thing to do :)

There is no overlap with the packages you mentioned:

- https://github.com/AndyBunn/dplR: this is for other purposes, mainly general analyses of tree-ring material. I am also a contributor to that package. In addition, my package uses that packages as a dependency.
- https://github.com/hanecakr/fellingdateR: I know that package and the author. That package serves another goal: mainly to analyse felling dates, while my package aims at making the use of network analysis accessible for dendrochronology
- https://github.com/ropensci/MtreeRing: This package is for measuring tree rings and not for visualising patterns in networks.
I believe that the dendroNetwork package can work very well with these packages together, but it adds new functionality to the available R-packages and I have tried to explain that in the paper (for JOSS) I have added in the repository.
Oh, for your information, there are more packages in R for dendrochronology and I have compiled a list in the past: https://ronaldvisser.github.io/Dendro_R/
