### 2024-03-13 #627 @kaijagahm
## Package Review

-   I have no relationship with the package author.
-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README

-   ☐ **Installation instructions:** for the development version of package and any non-standard dependencies in README

    Some functionality of the package depends on having Cytoscape installed. This is briefly mentioned in the README, but I would recommend being a little more explicit that this is an external program that will need to be installed and that the Cytoscape-related portions of the package won’t work without it. You might also mention that CyREST should be automatically installed with Cytoscape, and that if users encounter errors with CyREST they should try restarting their computer, maybe downloading CyREST separately, etc.

-   ☐ **Vignette(s):** demonstrating major functionality that runs successfully locally

    Vignettes exist, but `vignette(package = "dendroNetwork")` returns no results, `browseVignettes(package = "dendroNetwork")` returns no results, and the vignettes aren’t listed under `?dendroNetwork`. Make sure to include the vignettes when you build the package, and see <https://r-pkgs.org/vignettes.html> for reference on how to incorporate vignettes.

    Some of the vignettes (cytoscape.Rmd) do seem to run locally (and they are a bit clearer than the README about needing to install Cytoscape separately!) However, dendroNetwork\_use.Rmd fails to run locally because it needs citations… Maybe I’m just not understanding how to run a vignette properly? Maybe I need all the files there? Sorry if this is not a failure of the package but just a failure of me knowing how to run vignettes locally!

-   ☒ **Function Documentation:** for all exported functions

-   ☐ **Examples:** (that run successfully locally) for all exported functions

Example in `wuchswerte` fails because object `anos1` is not found.

    library(dendroNetwork)
    data(hol_rom)
    dplR::as.rwl(apply(anos1, 2, wuchswerte))
    #> Error in eval(expr, envir, enclos): object 'anos1' not found

`t_value` doesn’t return anything to console–consider including an explicit return statement.

    library(dendroNetwork)
    t_value(0.5, 100) # no output
    a <- t_value(0.5, 100)
    a
    #> [1] 5.744563

`cyto_create_gn_style(g_hol, gn_coms = g_hol_gn)` fails when run as part of the example for `cyto_create_gn_style`. Seems to be unable to load a file?

    library(dendroNetwork)
    data(hol_rom)
    sim_table_hol <- sim_table(hol_rom)
    g_hol <- dendro_network(sim_table_hol)
    hol_com_cpm_k3 <- clique_community_names(g_hol, k = 3)
    cyto_create_cpm_style(g_hol, k = 3, com_k = hol_com_cpm_k3)
    #> Error: Don't know how to read file:/private/var/folders/xy/lqbp515s27q8v1drkll9gynw0000gr/T/RtmppMCfZy/reprex-172671e49826-large-unau/

`cyto_create_cpm_style` fails similarly.

    library(dendroNetwork)
    data(hol_rom)
    sim_table_hol <- sim_table(hol_rom)
    g_hol <- dendro_network(sim_table_hol)
    hol_com_cpm_k3 <- clique_community_names(g_hol, k = 3)
    cyto_create_cpm_style(g_hol, k = 3, com_k = hol_com_cpm_k3)
    #> Error: Don't know how to read file:/private/var/folders/xy/lqbp515s27q8v1drkll9gynw0000gr/T/RtmppMCfZy/reprex-172625011c73-alive-lark/

-   ☐ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

Missing contribution guidelines. Includes URL, BugReports, and Maintainer.

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software been confirmed.
-   ☒ **Performance:** Any performance claims of the software been confirmed.
-   ☐ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

I see that there are tests set up, but I am not familiar with how to run the tests for someone else’s package on my own local machine. If anyone can provide guidance on this, I’d be happy to take a look! I have run tests for my own package before, but never for a package I’ve installed.

-   ☐ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Title: good

codemetar: I’m not familiar with this.

Platforms: runs on Mac; I can’t comment on other systems.

Package API: function names are in snake\_case and are reasonably logical. Could be nice to preface them all with a common thing like `dendro_` and make some of them shorter, but effort has already been made to make them short and consistent. I appreciate that the cytoscape-interface functions are all named with `cyto_`at the beginning, for instance.

Input checking: There is some input checking, and the error messages returned seem sensible. I would suggest doing a little more input checking, for example making sure that the network objects being passed to the functions are actually igraph network objects as expected. This goes hand in hand with my recommendation that you be a bit more specific about what data format is expected from users of the package. Keep in mind data types etc. and think about which ones will or won’t break your functions. I like to use the `checkmate` package for fast and easy argument checking, but there are many options including just more base R checks.

CITATION file: good

README etc: generally, multiple points of entry are done well. There is redundancy in where the information is presented, and I was mostly able to find out about the package and its core workflow through the README as well as the vignettes.

Documentation website: ?

Consider @family tag for function grouping

Estimated hours spent reviewing: 4

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

1.  I am coming to this review without any prior understanding of dendrochronology. Therefore, it would be helpful if you could provide more information about the data types. For example, `?hol_rom` tells me that this dataset contains dendrochronological site chronologies. I don’t actually know what a “site chronology” is. What is an object of class “rwl”, and how would I go about making sure that an object I’m feeding into this function has class “rwl” (or is it okay if it’s just a data frame)? What does a single row in this dataset represent? What does a single column represent? What do the column names mean? It’s possible that anyone who would be using this package would already understand these things, but I think there’s no harm in being more explicit. When I go to use a package, typically my first question is “If I wanted to use this package on my data, what would my data have to look like, exactly, in order for the functions to work?”

2.  Regarding the documentation for `sim_table`: I think I would understand this better if I understood the the data better, as described in 1). It seems like you’re saying that `trs1` (`hol_rom`, in this example) represents a single tree ring series (or is it multiple series? I can’t tell). `trs2` is optional. I did some experimenting and found that if you don’t specify `trs2`, the function ends up comparing `trs1` to itself; that is, `identical(sim_table(hol_rom), sim_table(hol_rom, hol_rom))` is `TRUE`. In the description for the `trs1` argument, maybe specify that behavior, instead of saying “to be compared with `trs2`.

3.  Carrying on with the theme of me being naive to the domain: I don’t understand what `last_digit_radius` does. What is a “series name” and what is the “radius of the tree-ring series”? What would be the units of the radius? Why does this argument matter? Again, if the answers to some of these questions would be obvious to domain-familiar users of your package, then maybe it’s not necessary to define them, but I was wondering.

4.  Make sure to specify in multiple places that Cytoscape has to be not only installed but actually running when you use the functions that relate to it.

5.  It would be nice if there was an example of how to use the package on large datasets. I recognize, though, that including a large dataset is probably not practical. Maybe use a small dataset as an example but still run it in parallel, using `tic()` and `toc()` to demonstrate the time difference? In the given example, it wasn’t clear that the object `network` in the example code wasn’t a real object and that the code wouldn’t actually run.

6.  `cyto_create_cpm_style` maybe didn’t work for me? I’m not exactly sure what was supposed to happen, but I didn’t see any colors or styling appearing, just black and white. I can provide more details/screenshots if that would help.

Some more detailed comments on the pre-cytoscape functions (some of this is maybe overly nitpicky, but it was my thought process as I was working through the functions!)
