### 2024-03-13 #627 @RonaldVisser
Hi @kaijagahm, thank you for the review and your detailed comments. These are very helpful for improving the package and the documentation, which makes me happy. I am looking forward to address the various things you pointed out and I hope to do this as soon as possible.
