### 2024-03-13 #627 @kaijagahm
You're very welcome! It was fun to explore this. I know my comments are a bit all over the place; I hope they are navigable. Feel free to ask if you encounter any confusion.

Except for some concerns over the examples running and one of the functions not showing colors in Cytoscape, and some documentation tweaks, most of my comments were around the margins and the package seems to be doing pretty well at its core functionality. I look forward to seeing how this develops!
