### 2024-04-04 #627 @gzach93
I think the author did a great job connecting their R package to the already existing R packages specifically created for their field.

The installation instructions were easy to follow and I was able to install the package and realted software needed.

It seems like the audience for this package would be new researchers to dendrology that are looking to create dendrochonological networks. If this is the intended audience maybe the author can add  a little more about what Cytoscape is and does or provide a small summary and where the reader can find more resources? I want to make sure that I understand but is Cytoscape mainly used for the visualization after this R packages formats the data and run the network analyses?

Could you add the data description for the example data in the vignettes? I see it in RING_Visser_2021.Rd, but it might be good to move that into the vignette.

Can more specific details about what the code is doing be added? Consider also adding a brief description of things like the Girvan-Newman algorithm or more technical terms. Again I am assuming that this package is trying to create an easy pathway for researchers getting into this sort of analysis.

The t_value function could use a return function at the end of it. The example of the t_value function does not output results right now.

The wuchswerte function example, I think, require a dataset "anos1" from the dplR package. However, in order to use this dataset users have to visit https://opendendro.github.io/dplR-workshop/ to get started. Is it possible to make this data more accessible or use another dataset?

I am not sure how to run these tests.

---

1) From my read through, I assumed that the audience was a new dendrologist of someone less familiar with creating these analyzes and this package creates an easier enterance point to help them. It think that this package does pretty good with this. However, maybe the vingettes can be expanded a little, offering another example dataset to work with and added definition for more field related jargon.

2) In the vignettes can you add a description of the example data. It might help new users better connect their datasets to the example data and think about the correct formatting needed.

3) I am not a dendrologist and I am not familiar with Cytoscape, so I am not sure if this is a familiar enough program where users do not need more information about what this program is used for or not. But, I think it might be useful to add more information about what Cytoscape is and some of the uses it has.
