### 2024-04-04 #627 @ldecicco-USGS
Thanks @gzach93! One way to run the tests (these are the "testthat" tests in the "test" folder) is to run `devtools::test()`. There are a few other options (you could go into the test folder and run them line-by-line even). The comment "Unit tests cover essential functions of the package and a reasonable range of inputs and conditions." means you are looking at the set of tests that the author includes to make sure they are testing a good-enough set of conditions on their functions. From the automated ropensci-bot, we know "Package coverage is 91.4%.". That's pretty good! It means 91.4% of the code is somehow tested in the testthat folder. The reviewer's job is just to make sure those tests seem adequate (which _is_ subjective, but you can generally get a pretty good idea if the author's doing their due-diligence). If you knew of some important edge-cases for example, you would maybe want the developers to be testing those conditions.

@RonaldVisser you can now start addressing the reviews!
