### 2024-04-05 #627 @RonaldVisser
## Response to review Zachary Gajewski (gzach93)

Thank you for your review, @gzach93. I am happy that you were positive about many things, but also for finding things to improve the package! Some things were already corrected before receiving your review and I am sorry for already updating while waiting for your review.

Documentation

Installation instructions: Thanks for noticing that. I have added a description of Cytoscape to the README and the vignettes: dendroNetwork and Cytoscape, including a link to the [Cytoscape-website](https://cytoscape.org/what_is_cytoscape.html)., see [commit](https://github.com/RonaldVisser/dendroNetwork/commit/bc1bd46c1b413350dd333e826503f72af9005b2f).

Vignettes:

I have added a vignette on dendrochronological data, see [commit](https://github.com/RonaldVisser/dendroNetwork/commit/aa29067c9000467df09f22d9ed4948697fc8d250).

Links to a better explanation of the Girvan-Newman Algorithm and CPM were already added earlier, based on the review by @kaijagahm, see [response and commit](https://github.com/ropensci/software-review/issues/627#issuecomment-2027514872)

Examples:

Thanks for also noticing that t_value() was not outputting to console. This was already corrected, based on the other review: see [commit](https://github.com/RonaldVisser/dendroNetwork/commit/d72bd4ca44a13628ca23e71389973013bd5c2743).

The references to the anos1 dataset was also already corrected, see [commit](https://github.com/RonaldVisser/dendroNetwork/commit/d72bd4ca44a13628ca23e71389973013bd5c2743)

 Automated tests: see [response](https://github.com/ropensci/software-review/issues/627#issuecomment-2027514872) to the review by [kaijagahm](https://github.com/kaijagahm)

You were added as reviewer to the DESCRIPTION-file, see [commit](https://github.com/RonaldVisser/dendroNetwork/commit/65d636b7e2cc77a17499ae4030f8e3cdd52f82a2)

Review Comments

Thanks for pointing this out. I have update various things in the readme and the vignettes. I hope that this is now more clear

I have added a vignette on the data, so this should help.

I have added some description for Cytoscape and also some links to find more help.

I would like to explain that a [dendrology](https://en.wikipedia.org/wiki/Dendrology) is not the same as [dendrochronology](https://en.wikipedia.org/wiki/Dendrochronology). Both are tree related, but the latter is aimed at tree-rings and the temporal patterns and the former more on trees in general. This confusion is very common, since these words look very similar.
