### 2024-04-09 #627 @ldecicco-USGS
Hi @gzach93 and @kaijagahm

Do you feel that the changes and updates have addressed your concerns?

If so, see:
https://devdevguide.netlify.app/approval2template
