### 2024-04-26 #627 @RonaldVisser
Thanks @ldecicco-USGS for your excellent role as editor! And thanks again @kaijagahm and @gzach93 for your wonderful reviews.

The package is now transferred to ROpenSci and I have addressed all to do's

I'd be happy to write a blog post about this packages @ropensci/blog-editors
