### 2024-05-02 #627 @RonaldVisser
Hi @steffilazerte ,

Thanks for the comments! I'll have a more through look at the documentation and I'll probably email you with questions later. I have not received a Slack-invite, but I have not used that for some time...

Thank you :)
